pyproject.toml
[project]
name = "news"
version = "0.1.0"
description = "REST API News"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.5",
    "django-cors-headers>=4.7.0",
    "django-filter>=25.1",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "pillow>=11.3.0",
    "psycopg2>=2.9.10",
    "psycopg2-binary>=2.9.10",
    "pyjwt>=2.10.1",
    "python-decouple>=3.8",
]
